backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jakub Piecha <piechajakub@wp.pl>"]
license = "Beerware"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.5"
pytest = "7.4.2"
pytest-django = "4.5.2"
pytest-cov = "4.1.0"
flake8 = "6.1.0"
black = "23.7.0"
flake8-django = "1.4"
django-sphinxdoc = "2.0.0"
djangorestframework = "3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]

addopts = [
    "--cov=.",
    "--cov-config=.coveragerc",
    "--cov-report",
    "term-missing:skip-covered",
    "--cov-fail-under=80",
]

[tool.pytest.markers]
slow = "Run slow unit test e.g. for models"
unit = "Run unit tests"
smoke = "Run core unit tests"
dbtest = "Run databasepoerty tests"


[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    | \git/ # exclude a directory
    | \.mypy_cache/ # exclude a directory
    | \.pytest_cache/ # exclude a directory
    | .*\.pyi? # excluded generated files
    | .*\.json # excluded netadata files
    | .*\.md # excluded documentation
    | .*\.toml # excluded configurationa files
    | .*\.yaml # excluded configurationa files
    | .*\.yml # excluded configurationa files
    | .*\.txt # excluded configurationa files
    | .*\.lock # excluded configurationa files
    | .*\.sh # excluded configurationa files
    | .*\.log # excluded configurationa files
)/
'''

